requirements.txt
psycopg2
jaydebeapi
pyyaml
polars
